ssa_alloc: drop redundant `idx as usize` cast in populate_param_ref_hints
`enumerate()` already yields `usize`; the cast tripped
clippy::unnecessary-cast under -D warnings.

+++ src/c5/codegen/ssa_alloc.rs
@@ -795,7 +795,7 @@ fn populate_param_ref_hints(func: &FunctionSsa, target: Target, hints: &mut [Opt
     for (idx, inst) in func.insts.iter().enumerate() {
         if let Inst::ParamRef { idx: i, .. } = inst
             && let Some(&r) = int_args.get(*i as usize)
-            && (idx as usize) < hints.len()
+            && idx < hints.len()
             && hints[idx].is_none()
         {
             hints[idx] = Some(r);
